+++ snet/route/type.go
@@ -1,0 +1,9 @@
+package route
+
+var(
+
+	meta_event = "meta_event"
+    message_event = "message"
+	notice_event = "notice"
+
+)
